BackEnd/requirements.txt
# ===========================================
# JanaSewa Backend Dependencies
# ===========================================
# Install: pip install -r requirements.txt
# ===========================================

# Web Framework
fastapi
uvicorn[standard]

# Data Validation
pydantic
pydantic-settings
email-validator

# Database
sqlalchemy
psycopg2-binary

# Authentication & Security
bcrypt
python-jose[cryptography]
python-multipart

# Configuration
python-dotenv
